Remove specification in workflow
There appears to be a conflict between what is stated in the workflow, and what is defined in the package.json.

The previous few commits are been due to pnpm not looking at the pnpm-lock.yaml file during the GitHub action, so these commits are essentially an investigation to try and get pnpm to play nice during the action.